finbot/requirements.txt
python-telegram-bot
sqlalchemy
apscheduler
google-genai
python-dotenv
